+++ lib/draggable_panel.dart
@@ -1,1 +1,2 @@
 export 'src/draggable_panel.dart';
+export 'src/controller/controller.dart';