+++ lib/src/draggable_panel.dart
@@ -1,10 +1,10 @@
 // ignore_for_file: inference_failure_on_function_return_type, prefer_int_literals, unnecessary_parenthesis, prefer_underscore_for_unused_callback_parameters, unnecessary_lambdas
 
+import 'package:draggable_panel/src/controller/controller.dart';
 import 'package:draggable_panel/src/enums/dock_type.dart';
 import 'package:draggable_panel/src/enums/panel_state.dart';
 import 'package:draggable_panel/src/utils/colors.dart';
 import 'package:draggable_panel/src/widgets/curve_line_paint.dart';
-import 'package:draggable_panel/src/widgets/multi_value_listenable.dart';
 import 'package:flutter/material.dart';
 part 'widgets/panel_button.dart';
 
@@ -24,15 +24,12 @@ class DraggablePanel extends StatefulWidget {
     this.buttonWidth = 35,
     this.buttonHeight = 70.0,
     this.borderRadius,
-    this.initialPanelState,
-    this.panelAnimDuration = 600,
-    this.dockType = DockType.inside,
-    this.dockOffset = 10.0,
-    this.dockAnimDuration = 300,
-    this.initialPosition,
+    this.controller,
     this.onPositionChanged,
     this.panelHeight,
   });
+
+  final DraggablePanelController? controller;
 
   /// The child widget that will be used as the main content of the screen.
   final Widget? child;
@@ -64,26 +61,6 @@ class DraggablePanel extends StatefulWidget {
   /// The background color of the panel.
   final Color? backgroundColor;
 
-  /// The initial state of the panel.
-  /// `PanelState.open` will open the panel when the widget is initialized.
-  /// `PanelState.closed` will close the panel when the widget is initialized.
-  final PanelState? initialPanelState;
-
-  /// The duration of the panel animation when it's opened or closed.
-  final int panelAnimDuration;
-
-  /// The type of the dock.
-  /// `DockType.inside` it will be located a little further from the edge of the screen.
-  /// `DockType.outside` it will be located a little closer to the middle of the screen.
-  final DockType? dockType;
-
-  /// The offset of the dock.
-  /// Default value is `10.0`.
-  final double dockOffset;
-
-  /// The duration of the dock animation.
-  final int? dockAnimDuration;
-
   /// The list of items that will be displayed in the panel.
   final List<
       ({
@@ -100,9 +77,6 @@ class DraggablePanel extends StatefulWidget {
         void Function(BuildContext context) onTap,
       })> buttons;
 
-  /// The initial position of the panel.
-  final ({double x, double y})? initialPosition;
-
   /// The callback that will be called when the position of the panel is changed.
   /// You can use local storage to save the position of the panel and restore it
   /// when the widget is initialized in `initialPosition`.
@@ -113,77 +87,32 @@ class DraggablePanel extends StatefulWidget {
 }
 
 class _DraggablePanelState extends State<DraggablePanel> {
-  // <-- Notifiers -->
-
-// Required to set the default state to closed when the widget gets initialized;
-  final ValueNotifier<PanelState> _panelState =
-      ValueNotifier(PanelState.closed);
-
-// Default positions for the panel;
-  final ValueNotifier<double> _draggablePositionTop = ValueNotifier(0.0);
-  final ValueNotifier<double> _draggablePositionLeft = ValueNotifier(0.0);
-  final ValueNotifier<double> _panelPositionLeft = ValueNotifier(0.0);
-
-  // ** PanOffset ** is used to calculate the distance from the edge of the panel
-  // to the cursor, to calculate the position when being dragged;
-
-  final ValueNotifier<double> _panOffsetTop = ValueNotifier(0.0);
-  final ValueNotifier<double> _panOffsetLeft = ValueNotifier(0.0);
-
-  // This is the animation duration for the panel movement, it's required to
-  // dynamically change the speed depending on what the panel is being used for.
-  // e.g: When panel opened or closed, the position should change in a different
-  // speed than when the panel is being dragged;
-
-  final ValueNotifier<int> _movementSpeed = ValueNotifier(0);
-
-  final ValueNotifier<bool> _isDragging = ValueNotifier(false);
-
-  final ValueNotifier<double> _buttonWidth = ValueNotifier(0.0);
-
-  static const double _panelWidth = 200;
+  late final DraggablePanelController _controller;
 
   @override
   void initState() {
     super.initState();
-    _draggablePositionTop.value = widget.initialPosition?.y ?? 200;
-    _draggablePositionLeft.value = widget.initialPosition?.x ?? 0;
-    _buttonWidth.value = widget.buttonWidth;
+    _controller = widget.controller ?? DraggablePanelController();
+    _controller.draggablePositionTop = _controller.initialPosition?.y ?? 200;
+    _controller.draggablePositionLeft = _controller.initialPosition?.x ?? 0;
+    _controller.buttonWidth = widget.buttonWidth;
 
-    _toggle();
-  }
-
-  @override
-  void didChangeDependencies() {
-    _toggle();
-    _isDragging.value = false;
-    _panelState.value = PanelState.closed;
-    super.didChangeDependencies();
-  }
-
-  void _toggle() {
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (widget.initialPanelState == PanelState.open) {
-        _toggleMainButton(MediaQuery.sizeOf(context).width);
-        _togglePanel(MediaQuery.sizeOf(context).width);
-      } else {
-        _forceDock(MediaQuery.sizeOf(context).width);
-        _hidePanel(MediaQuery.sizeOf(context).width);
-      }
+      _controller.toggle(context);
     });
   }
 
   @override
+  void didChangeDependencies() {
+    _controller.toggle(context);
+    _controller.isDragging = false;
+    _controller.panelState = PanelState.closed;
+    super.didChangeDependencies();
+  }
+
+  @override
   void dispose() {
-    _panelState.dispose();
-    _draggablePositionTop.dispose();
-    _draggablePositionLeft.dispose();
-    _panOffsetTop.dispose();
-    _panOffsetLeft.dispose();
-    _movementSpeed.dispose();
-    _isDragging.dispose();
-    _buttonWidth.dispose();
-    _panelPositionLeft.dispose();
+    _controller.dispose();
     super.dispose();
   }
 
@@ -193,99 +122,90 @@ class _DraggablePanelState extends State<DraggablePanel> {
     final pageWidth = MediaQuery.sizeOf(context).width;
     final pageHeight = MediaQuery.sizeOf(context).height;
 
-    return MultiValueListenableBuilder(
+    return AnimatedBuilder(
       key: const ValueKey('draggable_panel_builder'),
-      valueListenables: [
-        _panelState,
-        _draggablePositionTop,
-        _draggablePositionLeft,
-        _panelPositionLeft,
-        _movementSpeed,
-        _isDragging,
-        _buttonWidth,
-        _panOffsetTop,
-        _panOffsetLeft,
-      ],
-      builder: (context) {
+      animation: _controller,
+      builder: (context, child) {
         // Animated positioned widget can be moved to any part of the screen with
         // animation;
-        final isInRightSide = _draggablePositionLeft.value > pageWidth / 2;
+        final isInRightSide = _controller.draggablePositionLeft > pageWidth / 2;
         return Stack(
           children: [
             if (widget.child != null) widget.child!,
             AnimatedPositioned(
               key: const ValueKey('draggable_panel_button'),
               duration: Duration(
-                milliseconds: _movementSpeed.value,
+                milliseconds: _controller.movementSpeed,
               ),
-              top: _draggablePositionTop.value,
-              left: _draggablePositionLeft.value,
+              top: _controller.draggablePositionTop,
+              left: _controller.draggablePositionLeft,
               curve: Curves.fastLinearToSlowEaseIn,
               child: GestureDetector(
                 onPanEnd: (event) {
-                  _isDragging.value = false;
-                  _forceDock(pageWidth);
-                  _hidePanel(pageWidth);
+                  _controller.isDragging = false;
+                  _controller.forceDock(pageWidth);
+                  _controller.hidePanel(pageWidth);
                   widget.onPositionChanged?.call(
-                    _draggablePositionLeft.value,
-                    _draggablePositionTop.value,
+                    _controller.draggablePositionLeft,
+                    _controller.draggablePositionTop,
                   );
                 },
                 onPanStart: (event) {
                   // Detect the offset between the top and left side of the panel and
                   // x and y position of the touch(click) event;
-                  _panOffsetTop.value =
-                      event.globalPosition.dy - _draggablePositionTop.value;
-                  _panOffsetLeft.value =
-                      event.globalPosition.dx - _draggablePositionLeft.value;
+                  _controller.panOffsetTop = event.globalPosition.dy -
+                      _controller.draggablePositionTop;
+                  _controller.panOffsetLeft = event.globalPosition.dx -
+                      _controller.draggablePositionLeft;
                 },
                 onPanUpdate: (event) {
                   // Close Panel if opened;
-                  _panelState.value = PanelState.closed;
+                  _controller.panelState = PanelState.closed;
 
-                  _buttonWidth.value = widget.buttonWidth;
+                  _controller.buttonWidth = widget.buttonWidth;
 
                   // Reset Movement Speed;
-                  _movementSpeed.value = 0;
-                  _isDragging.value = true;
+                  _controller.movementSpeed = 0;
+                  _controller.isDragging = true;
 
                   // Calculate the top position of the panel according to pan;
                   final statusBarHeight = MediaQuery.paddingOf(context).top;
-                  _draggablePositionTop.value =
-                      event.globalPosition.dy - _panOffsetTop.value;
+                  _controller.draggablePositionTop =
+                      event.globalPosition.dy - _controller.panOffsetTop;
 
                   // Check if the top position is exceeding the status bar or dock boundaries;
-                  if (_draggablePositionTop.value <
+                  if (_controller.draggablePositionTop <
                       statusBarHeight + _dockBoundary) {
-                    _draggablePositionTop.value =
+                    _controller.draggablePositionTop =
                         statusBarHeight + _dockBoundary;
                   }
-                  if (_draggablePositionTop.value >
+                  if (_controller.draggablePositionTop >
                       (pageHeight - widget.buttonHeight - 10) - _dockBoundary) {
-                    _draggablePositionTop.value =
+                    _controller.draggablePositionTop =
                         (pageHeight - widget.buttonHeight - 10) - _dockBoundary;
                   }
 
                   // Calculate the Left position of the panel according to pan;
-                  _draggablePositionLeft.value =
-                      event.globalPosition.dx - _panOffsetLeft.value;
+                  _controller.draggablePositionLeft =
+                      event.globalPosition.dx - _controller.panOffsetLeft;
 
                   // Check if the left position is exceeding the dock boundaries;
-                  if (_draggablePositionLeft.value < 0 + _dockBoundary) {
-                    _draggablePositionLeft.value = 0 + _dockBoundary;
+                  if (_controller.draggablePositionLeft < 0 + _dockBoundary) {
+                    _controller.draggablePositionLeft = 0 + _dockBoundary;
                   }
-                  if (_draggablePositionLeft.value >
-                      (pageWidth - _buttonWidth.value) - _dockBoundary) {
-                    _draggablePositionLeft.value =
-                        (pageWidth - _buttonWidth.value) - _dockBoundary;
+                  if (_controller.draggablePositionLeft >
+                      (pageWidth - _controller.buttonWidth) - _dockBoundary) {
+                    _controller.draggablePositionLeft =
+                        (pageWidth - _controller.buttonWidth) - _dockBoundary;
                   }
                 },
                 onTap: () async {
-                  await _toggleMainButton(pageWidth);
-                  _togglePanel(pageWidth);
+                  await _controller.toggleMainButton(pageWidth);
+                  _controller.togglePanel(pageWidth);
                 },
                 child: AnimatedContainer(
-                  duration: Duration(milliseconds: widget.panelAnimDuration),
+                  duration:
+                      Duration(milliseconds: _controller.panelAnimDuration),
                   width: widget.buttonWidth,
                   height: widget.buttonHeight,
                   clipBehavior: Clip.antiAlias,
@@ -305,49 +225,46 @@ class _DraggablePanelState extends State<DraggablePanel> {
                       children: [
                         // Gesture detector is required to detect the tap and drag on the panel;
                         Flexible(
-                          child: _panelState.value == PanelState.open
-                              ? const SizedBox()
-                              : AnimatedSwitcher(
-                                  duration: const Duration(milliseconds: 200),
-                                  transitionBuilder: (child, animation) =>
-                                      ScaleTransition(
-                                    scale: animation,
-                                    child: child,
-                                  ),
-                                  child: _isDragging.value
-                                      ? Center(
-                                          child: SizedBox(
-                                            width: _buttonWidth.value,
-                                            height: widget.buttonHeight,
-                                            child: Icon(
-                                              Icons.drag_indicator_rounded,
+                          child: AnimatedSwitcher(
+                            duration: const Duration(milliseconds: 200),
+                            transitionBuilder: (child, animation) =>
+                                ScaleTransition(
+                              scale: animation,
+                              child: child,
+                            ),
+                            child: _controller.isDragging
+                                ? Center(
+                                    child: SizedBox(
+                                      width: _controller.buttonWidth,
+                                      height: widget.buttonHeight,
+                                      child: Icon(
+                                        Icons.drag_indicator_rounded,
+                                        color:
+                                            Colors.white.withValues(alpha: 0.5),
+                                      ),
+                                    ),
+                                  )
+                                : SizedBox(
+                                    key: const ValueKey('drag_handle'),
+                                    width: _controller.buttonWidth,
+                                    height: widget.buttonHeight,
+                                    child: widget.icon ??
+                                        Align(
+                                          alignment: isInRightSide
+                                              ? Alignment.centerLeft
+                                              : Alignment.centerRight,
+                                          child: CustomPaint(
+                                            willChange: true,
+                                            size: const Size(20, 65),
+                                            painter: LineWithCurvePainter(
+                                              isInRightSide: isInRightSide,
                                               color: Colors.white
                                                   .withValues(alpha: 0.5),
                                             ),
                                           ),
-                                        )
-                                      : SizedBox(
-                                          key: const ValueKey('drag_handle'),
-                                          width: _buttonWidth.value,
-                                          height: widget.buttonHeight,
-                                          child: widget.icon ??
-                                              Align(
-                                                alignment: isInRightSide
-                                                    ? Alignment.centerLeft
-                                                    : Alignment.centerRight,
-                                                child: CustomPaint(
-                                                  willChange: true,
-                                                  size: const Size(20, 65),
-                                                  painter: LineWithCurvePainter(
-                                                    isInRightSide:
-                                                        isInRightSide,
-                                                    color: Colors.white
-                                                        .withValues(alpha: 0.5),
-                                                  ),
-                                                ),
-                                              ),
                                         ),
-                                ),
+                                  ),
+                          ),
                         ),
                       ],
                     ),
@@ -358,33 +275,34 @@ class _DraggablePanelState extends State<DraggablePanel> {
             AnimatedPositioned(
               key: const ValueKey('draggable_panel'),
               duration: Duration(
-                milliseconds: widget.panelAnimDuration,
+                milliseconds: _controller.panelAnimDuration,
               ),
               top: _panelTopPosition(pageHeight),
-              left: _panelPositionLeft.value,
+              left: _controller.panelPositionLeft,
               curve: Curves.linearToEaseOut,
               child: TapRegion(
                 onTapOutside: (event) {
-                  if (_panelState.value == PanelState.open) {
-                    _panelState.value = PanelState.closed;
+                  if (_controller.panelState == PanelState.open) {
+                    _controller.panelState = PanelState.closed;
 
                     // Reset panel position, dock it to nearest edge;
-                    _forceDock(pageWidth);
-                    _togglePanel(pageWidth);
+                    _controller.forceDock(pageWidth);
+                    _controller.togglePanel(pageWidth);
                   }
                 },
                 child: AnimatedSwitcher(
-                  duration: Duration(milliseconds: widget.panelAnimDuration),
+                  duration:
+                      Duration(milliseconds: _controller.panelAnimDuration),
                   transitionBuilder: (child, animation) => Transform.translate(
                     offset: Offset.zero,
                     child: child,
                   ),
-                  child: _panelState.value == PanelState.open
+                  child: _controller.panelState == PanelState.open
                       ? AnimatedContainer(
                           key: const ValueKey('panel_container'),
-                          duration:
-                              Duration(milliseconds: widget.panelAnimDuration),
-                          width: _panelWidth,
+                          duration: Duration(
+                              milliseconds: _controller.panelAnimDuration),
+                          width: _controller.panelWidth,
                           height: _panelHeight,
                           clipBehavior: Clip.antiAlias,
                           decoration: BoxDecoration(
@@ -436,10 +354,10 @@ class _DraggablePanelState extends State<DraggablePanel> {
                                             widget.items[index].onTap
                                                 .call(context);
 
-                                            _panelState.value =
+                                            _controller.panelState =
                                                 PanelState.closed;
-                                            _forceDock(pageWidth);
-                                            _hidePanel(pageWidth);
+                                            _controller.forceDock(pageWidth);
+                                            _controller.hidePanel(pageWidth);
                                           },
                                         ),
                                       ),
@@ -482,66 +400,24 @@ class _DraggablePanelState extends State<DraggablePanel> {
   // <-- Functions -->
 
   double _panelTopPosition(double pageHeight) {
-    if (_draggablePositionTop.value < 0) {
+    if (_controller.draggablePositionTop < 0) {
       return 0;
-    } else if (_draggablePositionTop.value > pageHeight - _panelHeight) {
-      return _draggablePositionTop.value - _panelHeight;
+    } else if (_controller.draggablePositionTop > pageHeight - _panelHeight) {
+      return _controller.draggablePositionTop - _panelHeight;
     } else {
-      return _draggablePositionTop.value + widget.buttonHeight;
-    }
-  }
-
-  void _hidePanel(double pageWidth) {
-    if (_draggablePositionLeft.value > pageWidth / 2) {
-      _panelPositionLeft.value = pageWidth + _panelWidth;
-    } else {
-      _panelPositionLeft.value = -_panelWidth;
-    }
-  }
-
-  void _togglePanel(double pageWidth) {
-    if (_panelState.value == PanelState.open) {
-      if (_draggablePositionLeft.value > pageWidth / 2) {
-        _panelPositionLeft.value = pageWidth - _panelWidth - widget.buttonWidth;
-      } else {
-        _panelPositionLeft.value = widget.buttonWidth;
-      }
-    } else {
-      if (_draggablePositionLeft.value > pageWidth / 2) {
-        _panelPositionLeft.value = pageWidth;
-      } else {
-        _panelPositionLeft.value = -_panelWidth;
-      }
-    }
-  }
-
-  Future<void> _toggleMainButton(double pageWidth) async {
-    if (_panelState.value == PanelState.open) {
-      // If panel state is "open", set it to "closed";
-      _panelState.value = PanelState.closed;
-
-      // Reset panel position, dock it to nearest edge;
-      _forceDock(pageWidth);
-    } else {
-      // If panel state is "closed", set it to "open";
-      _panelState.value = PanelState.open;
-
-      if (_draggablePositionLeft.value > pageWidth / 2) {
-        _draggablePositionLeft.value = pageWidth + _buttonWidth.value;
-      } else {
-        _draggablePositionLeft.value = -_buttonWidth.value;
-      }
+      return _controller.draggablePositionTop + widget.buttonHeight;
     }
   }
 
   double get _dockBoundary {
-    if (widget.dockType != null && widget.dockType == DockType.inside) {
+    if (_controller.dockType != null &&
+        _controller.dockType == DockType.inside) {
       // If it's an 'inside' type dock, dock offset will remain the same;
-      return -widget.dockOffset;
+      return -_controller.dockOffset;
     } else {
       // If it's an 'outside' type dock, dock offset will be inverted, hence
       // negative value;
-      return widget.dockOffset;
+      return _controller.dockOffset;
     }
   }
 
@@ -582,30 +458,6 @@ class _DraggablePanelState extends State<DraggablePanel> {
       );
     } else {
       return null;
-    }
-  }
-
-  // Force dock will dock the panel to it's nearest edge of the screen;
-  void _forceDock(double pageWidth) {
-    // Calculate the center of the panel;
-    final center = _draggablePositionLeft.value + (_buttonWidth.value / 2);
-
-    // Set movement speed to the custom duration property or '300' default;
-    _movementSpeed.value = widget.dockAnimDuration ?? 300;
-
-    _buttonWidth.value = widget.buttonWidth;
-
-    // Check if the position of center of the panel is less than half of the
-    // page;
-    if (center < pageWidth / 2) {
-      // Dock to the left edge;
-      _draggablePositionLeft.value = 0.0 + _dockBoundary;
-      _panelPositionLeft.value = -_buttonWidth.value;
-    } else {
-      // Dock to the right edge;
-      _draggablePositionLeft.value =
-          (pageWidth - _buttonWidth.value) - _dockBoundary;
-      _panelPositionLeft.value = pageWidth + _buttonWidth.value;
     }
   }